+++ library/SolverUtils/Filters/FilterMovingAverage.cpp
@@ -61,7 +61,8 @@ FilterMovingAverage::FilterMovingAverage(
         else
         {
             // Load time constant
-            LibUtilities::Equation equ(m_session, it->second);
+            LibUtilities::Equation equ(
+                m_session->GetExpressionEvaluator(), it->second);
             NekDouble tau = equ.Evaluate();
             // Load delta T between samples
             NekDouble dT;
@@ -73,7 +74,8 @@ FilterMovingAverage::FilterMovingAverage(
     }
     else
     {
-        LibUtilities::Equation equ(m_session, it->second);
+        LibUtilities::Equation equ(
+            m_session->GetExpressionEvaluator(), it->second);
         m_alpha = equ.Evaluate();
         // Check if tau was also defined
         it = pParams.find("tau");