+++ library/Timings/TimingCGHelmSolve3D.cpp
@@ -3,7 +3,7 @@
 #include <iomanip>
 
 #include <boost/filesystem/path.hpp>
-#include <SpatialDomains/MeshGraph3D.h>
+#include <SpatialDomains/MeshGraph.h>
 #include <MultiRegions/ContField3D.h>
 
 using namespace std;
@@ -201,7 +201,7 @@ int main(int argc, char *argv[])
     
     //----------------------------------------------
     // Read in mesh from input file
-    SpatialDomains::MeshGraphSharedPtr graph3D = MemoryManager<SpatialDomains::MeshGraph3D>::AllocateSharedPtr(vSession);
+    SpatialDomains::MeshGraphSharedPtr graph3D = SpatialDomains::MeshGraph::Read(vSession);;
     //----------------------------------------------
 
     //----------------------------------------------