+++ solvers/DiffusionSolver/DiffusionSolverTimeInt.cpp
@@ -86,6 +86,9 @@ Diffusion::Diffusion(int argc, char* argv[])
     // Create session reader.
     session     = LibUtilities::SessionReader::CreateInstance(argc, argv);
 
+    // Read the geometry and the expansion information
+    graph       = SpatialDomains::MeshGraph::Read(session);
+
     // Create Field I/O object.
     fld         = LibUtilities::FieldIO::CreateDefault(session);
 
@@ -96,9 +99,6 @@ Diffusion::Diffusion(int argc, char* argv[])
     delta_t     = session->GetParameter("TimeStep");
     epsilon     = session->GetParameter("epsilon");
     lambda      = 1.0/delta_t/epsilon;
-
-    // Read the geometry and the expansion information
-    graph       = SpatialDomains::MeshGraph::Read(session);
 
     // Set up the field
     field       = MemoryManager<MultiRegions::ContField2D>::