+++ solvers/IncNavierStokesSolver/EquationSystems/IncNavierStokes.cpp
@@ -67,9 +67,10 @@ namespace Nektar
      * \param
      */
     IncNavierStokes::IncNavierStokes(
-        const LibUtilities::SessionReaderSharedPtr& pSession):
-        UnsteadySystem(pSession),
-        AdvectionSystem(pSession),
+        const LibUtilities::SessionReaderSharedPtr& pSession,
+        const SpatialDomains::MeshGraphSharedPtr &pGraph):
+        UnsteadySystem(pSession, pGraph),
+        AdvectionSystem(pSession, pGraph),
         m_SmoothAdvection(false)
     {
     }