+++ solvers/IncNavierStokesSolver/EquationSystems/VelocityCorrectionScheme.h
@@ -46,10 +46,12 @@ namespace Nektar
 
         /// Creates an instance of this class
         static SolverUtils::EquationSystemSharedPtr create(
-                const LibUtilities::SessionReaderSharedPtr& pSession) {
+            const LibUtilities::SessionReaderSharedPtr& pSession,
+            const SpatialDomains::MeshGraphSharedPtr &pGraph)
+        {
             SolverUtils::EquationSystemSharedPtr p =
-                                MemoryManager<VelocityCorrectionScheme>::
-                                            AllocateSharedPtr(pSession);
+                MemoryManager<VelocityCorrectionScheme>::AllocateSharedPtr(
+                    pSession, pGraph);
             p->InitObject();
             return p;
         }
@@ -58,7 +60,9 @@ namespace Nektar
         static std::string className;
 
         /// Constructor.
-        VelocityCorrectionScheme(const LibUtilities::SessionReaderSharedPtr& pSession);
+        VelocityCorrectionScheme(
+            const LibUtilities::SessionReaderSharedPtr& pSession,
+            const SpatialDomains::MeshGraphSharedPtr &pGraph);
 
         virtual ~VelocityCorrectionScheme();